add admin login

+++ server.js
@@ -8,6 +8,7 @@ require("dotenv").config();
 const PORT = process.env.PORT || 5000;
 const app = express();
 const prisma = new PrismaClient();
+const adminPassword = process.env.ADMIN_PASSWORD;
 
 app.use(cors()); 
 app.use(express.json()); 
@@ -22,7 +23,7 @@ const checkAdmin = (req, res, next) => {
   }
 };
 
-// ROUTE: Get all published articles
+// PUBLIC ROUTE: Get all published articles
 app.get('/api/articles', async (req, res) => {
   const articles = await prisma.article.findMany({
     where: { published: true }
@@ -30,7 +31,7 @@ app.get('/api/articles', async (req, res) => {
   res.json(articles);
 });
 
-// GET one article by slug
+// PUBLIC GET one article by slug
 app.get('/api/articles/:slug', async (req, res) => {
   const { slug } = req.params;
   const article = await prisma.article.findUnique({
@@ -40,9 +41,26 @@ app.get('/api/articles/:slug', async (req, res) => {
 });
 
 
-// ROUTE: Create a new article
+// ADMIN LOGIN
+app.post('/api/login', (req, res) => {
+  const { password } = req.body;
+  if (password === adminPassword) {
+    // Add JWT token in future
+    res.json({ success: true, token: "secret-admin-token" });
+  } else {
+    res.status(401).json({ success: false, message: "Invalid Password" });
+  }
+});
+
+// ADMIN CREATE a new article
 app.post('/api/articles', checkAdmin, async (req, res) => {
   const { title, content } = req.body; 
+
+  const token = req.headers['authorization'];
+  
+  if (token !== "secret-admin-token") {
+    return res.status(403).json({ error: "Forbidden: You are not the admin" });
+  }
 
   const generatedSlug = slugify(title, {
     lower: true,  
@@ -65,10 +83,16 @@ app.post('/api/articles', checkAdmin, async (req, res) => {
 });
 
 
-// UPDATE an article by ID
+// ADMIN UPDATE an article by ID
 app.put('/api/articles/:id', checkAdmin, async (req, res) => {
   const { id } = req.params;
   const { title, content } = req.body;
+
+  const token = req.headers['authorization'];
+  
+  if (token !== "secret-admin-token") {
+    return res.status(403).json({ error: "Forbidden: You are not the admin" });
+  }
 
   try {
     const updatedArticle = await prisma.article.update({
@@ -81,8 +105,16 @@ app.put('/api/articles/:id', checkAdmin, async (req, res) => {
   }
 });
 
+// ADMIN DELETE an article by ID
 app.delete('/api/articles/:id', checkAdmin,async (req, res) => {
   const { id } = req.params;
+
+  const token = req.headers['authorization'];
+  
+  if (token !== "secret-admin-token") {
+    return res.status(403).json({ error: "Forbidden: You are not the admin" });
+  }
+  
   try {
     await prisma.article.delete({
       where: { id: parseInt(id) },
